backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0

# LangGraph and LangChain - Compatible versions
langgraph==0.0.62
langchain==0.2.16
langchain-community==0.2.16
langchain-aws==0.1.17

# AWS Bedrock - Compatible with langchain-aws
boto3>=1.34.131,<1.35.0
botocore>=1.34.131,<1.35.0

# Vector Database
chromadb==0.4.24
sentence-transformers==2.7.0

# Web scraping and content extraction
requests==2.31.0
beautifulsoup4==4.12.3
newspaper3k==0.2.8
lxml[html_clean]==4.9.3

# Additional utilities
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
